Added the functionality that only Admin can delete user

+++ quora-service/src/main/java/com/upgrad/quora/service/business/UserAdminBusinessService.java
@@ -1,11 +1,9 @@
 package com.upgrad.quora.service.business;
 
 import com.upgrad.quora.service.dao.UserDao;
-import com.upgrad.quora.service.entity.UserEntity;
+import com.upgrad.quora.service.entity.UserAuthTokenEntity;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Propagation;
-
 import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -16,8 +14,18 @@ public class UserAdminBusinessService {
     private UserDao userDao;
 
     @Transactional(propagation = Propagation.REQUIRED)
-    public void deleteUser(final String userUuid) {
-        userDao.deleteUser(userUuid);
+    public void deleteUser(final String userUuid, final String authorization) {
+
+        UserAuthTokenEntity userAuthTokenEntity = userDao.getUserAuthToken(authorization);
+
+        String role = userAuthTokenEntity.getUser().getRole();
+
+        System.out.println("The role is " + role + " auth " + authorization);
+        if(role.equals("admin")) {
+
+            userDao.deleteUser(userUuid);
+
+        }
 
     }
 }
